@@ -1,7 +1,5 @@
-from setuptools import setup, find_packages
-from setuptools.command.test import test as TestCommand
 import os
-import sys
+from setuptools import setup, find_packages
 
 
 def get_version():
@@ -11,6 +9,7 @@ def get_version():
         exec(f.read(), variables)
         return variables.get('VERSION')
     raise RuntimeError('No version info found.')
+
 
 setup(
     name='textile',
@@ -29,18 +28,18 @@ setup(
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3 :: Only',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
+        'Programming Language :: Python :: 3.12',
         'Topic :: Software Development :: Libraries :: Python Modules',
     ],
     keywords='textile,text,html markup',
     install_requires=[
         'html5lib>=1.0.1',
         'regex>1.0; implementation_name != "pypy"',
-        ],
+    ],
     extras_require={
         'develop': ['pytest', 'pytest-cov'],
         'imagesize': ['Pillow>=3.0.0'],
@@ -49,5 +48,5 @@ setup(
     tests_require=['pytest', 'pytest-cov'],
     include_package_data=True,
     zip_safe=False,
-    python_requires='>=3.5',
+    python_requires='>=3.8',
 )
