@@ -1,10 +1,11 @@
 from textile import Textile
 
+
 def test_span():
     t = Textile()
     result = t.retrieveTags(t.span("hello %(bob)span *strong* and **bold**% goodbye"))
     expect = ('hello <span class="bob">span <strong>strong</strong> and '
-            '<b>bold</b></span> goodbye')
+              '<b>bold</b></span> goodbye')
     assert result == expect
 
     result = t.retrieveTags(t.span('%:http://domain.tld test%'))
