@@ -33,5 +33,5 @@ def main():
         outfile.write(output)
 
 
-if __name__ == '__main__': #pragma: no cover
+if __name__ == '__main__':  # pragma: no cover
     main()
